pyproject.toml
[project]
name = "pizza-box"
version = "0.1.0"
description = "rest api for pizza-box pizzeria"
readme = "README.md"
requires-python = ">=3.13.5"
dependencies = [
    "alembic>=1.16.4",
    "asgiref>=3.9.1",
    "asyncpg>=0.30.0",
    "boto3>=1.40.18",
    "celery>=5.5.3",
    "fastapi-mail>=1.5.0",
    "fastapi[standard]>=0.116.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "razorpay>=2.0.0",
    "redis>=6.4.0",
    "sqlalchemy[asyncio]>=2.0.42",
    "types-passlib>=1.7.7.20250602",
    "websockets>=15.0.1",
]
[tool.fastapi]
entrypoint="app.main:app"
